Move 'get_device_properties()' method to BaseTV class
